Création d'un fichier pour stocker les données en dur
