+++ react-page-web/src/data/ReactThemesList.tsx
@@ -1,0 +1,56 @@
+import ReactThemes from "../components/models/ReactThemes";
+
+const reactThemesList: ReactThemes[] = [
+  {
+    id: 1,
+    title: "Modernize React MUI Dashboard Template",
+    description: "",
+    tag: "",
+    price: 59.00,
+    sales: 0,
+    downloads: 0,
+    notation: 0,
+  },
+  {
+    id: 2,
+    title: "Modernize Free React MUI Dashboard Template",
+    description: "",
+    tag: "",
+    price: 0.00,
+    sales: 0,
+    downloads: 0,
+    notation: 0,
+  },
+  {
+    id: 3,
+    title: "MaterialPro React Redux Admin",
+    description: "React Redux-Toolkit Admin",
+    tag: "React",
+    price: 49.00,
+    sales: 265,
+    downloads: 0,
+    notation: 0,
+  },
+  {
+    id: 4,
+    title: "Flexy - Bootstrap 5, React, Next js, Vuetify & Angular Dashboard",
+    description: "Production Ready Admin Template",
+    tag: "",
+    price: 59.00,
+    sales: 1105,
+    downloads: 0,
+    notation: 4.78,
+  },
+  {
+    id: 5,
+    title: "Xtreme React Admin Lite",
+    description: "Free React Admin Template with hooks",
+    tag: "React",
+    price: 0,
+    sales: 0,
+    downloads: 5817,
+    notation: 0,
+  },
+];
+
+export default reactThemesList;
